Release
```
           Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   14.58%  642.29ms      5000  128.46us  6.1120us  340.77us  kernel_find_ntuplets(TrackingRecHit2DSOAView const *, GPUCACell*, unsigned int const *, cms::cuda::SimpleVector<cms$
                   13.03%  573.68ms      5000  114.74us  16.704us  618.75us  gpuClustering::findClus(unsigned short const *, unsigned short const *, unsigned short const *, unsigned int const $
                    9.25%  407.24ms     10000  40.724us  2.5600us  204.64us  void kernel_BLFit<int=4>(cms::cuda::OneToManyAssoc<unsigned short, int=8, int=24576> const *, double, TrackSoAHeter$
                    8.82%  388.30ms      5000  77.659us  6.5590us  245.57us  gpuVertexFinder::vertexFinderOneKernel(ZVertexSoA*, gpuVertexFinder::WorkSpace*, int, float, float, float)
                    7.89%  347.50ms      5000  69.499us  9.1200us  316.80us  gpuPixelDoublets::getDoubletsFromHisto(GPUCACell*, unsigned int*, cms::cuda::SimpleVector<cms::cuda::VecArray<unsig$
                    5.09%  223.98ms      5000  44.795us  3.3280us  209.54us  gpuPixelDoublets::fishbone(TrackingRecHit2DSOAView const *, GPUCACell*, unsigned int const *, cms::cuda::VecArray<u$
                    5.03%  221.56ms      5000  44.311us  20.544us  174.02us  kernel_connect(cms::cuda::AtomicPairCounter*, cms::cuda::AtomicPairCounter*, TrackingRecHit2DSOAView const *, GPUCA$
                    3.99%  175.88ms      5000  35.176us  9.8230us  161.92us  gpuPixelRecHits::getHits(pixelCPEforGPU::ParamsOnGPU const *, BeamSpotPOD const *, SiPixelDigisCUDA::DeviceConstVie$
                    3.96%  174.45ms     35010  4.9820us  1.0240us  635.20us  [CUDA memcpy HtoD]
```

This PR

```
          Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   13.63%  558.73ms      5000  111.75us  15.072us  663.13us  gpuClustering::findClus(unsigned short const *, unsigned short const *, unsigned short const *, unsigned int const $
                    9.98%  409.10ms      5000  81.820us  6.0480us  193.79us  kernel_find_ntuplets(TrackingRecHit2DSOAView const *, GPUCACell*, unsigned int const *, cms::cuda::SimpleVector<cms$
                    9.86%  404.24ms     10000  40.424us  2.5600us  111.68us  void kernel_BLFit<int=4>(cms::cuda::OneToManyAssoc<unsigned short, int=8, int=24576> const *, double, TrackSoAHeter$
                    9.41%  385.77ms      5000  77.153us  6.7520us  146.88us  gpuVertexFinder::vertexFinderOneKernel(ZVertexSoA*, gpuVertexFinder::WorkSpace*, int, float, float, float)
                    8.32%  341.18ms      5000  68.235us  8.8960us  211.33us  gpuPixelDoublets::getDoubletsFromHisto(GPUCACell*, unsigned int*, cms::cuda::SimpleVector<cms::cuda::VecArray<unsig$
                    5.43%  222.45ms      5000  44.490us  3.3280us  119.14us  gpuPixelDoublets::fishbone(TrackingRecHit2DSOAView const *, GPUCACell*, unsigned int const *, cms::cuda::VecArray<u$
                    5.36%  219.69ms      5000  43.937us  20.480us  168.96us  kernel_connect(cms::cuda::AtomicPairCounter*, cms::cuda::AtomicPairCounter*, TrackingRecHit2DSOAView const *, GPUCA$
                    4.16%  170.61ms      5000  34.121us  10.367us  97.888us  gpuPixelRecHits::getHits(pixelCPEforGPU::ParamsOnGPU const *, BeamSpotPOD const *, SiPixelDigisCUDA::DeviceConstVie$
                    4.16%  170.50ms     35010  4.8700us  1.0240us  595.10us  [CUDA memcpy HtoD]
                    3.32%  135.98ms     15000  9.0650us  4.5760us  54.048us  void cms::cuda::multiBlockPrefixScan<unsigned int>(unsigned int const *, cms::cuda::multiBlockPrefixScan<unsigned i$
```


Throughput (triplets on T4)

release
  702.0 ±   1.6 ev/s
this PR
   752.4 ±   1.4 ev/s
only the first commit
  750.5 ±   1.2 ev/s

Purely technical. no regression expected. no regression observed