@cmsbuild 2021-06-25
A new Pull Request was created by  @VinInn (Vincenzo Innocente) for master.

It involves the following packages:

HeterogeneousCore/CUDAUtilities
RecoLocalTracker/SiPixelClusterizer
RecoPixelVertexing/PixelTriplets
RecoPixelVertexing/PixelVertexFinding


@perrotta, @makortel, @fwyzard, @jpata, @slava77 can you please review it and eventually sign? Thanks.
@mtosi, @makortel, @felicepantaleo, @GiacomoSguazzoni, @JanFSchulte, @rovere, @VinInn, @OzAmram, @ferencek, @dkotlins, @gpetruc, @threus, @dgulhan, @tvami this is something you requested to watch as well.
@silviodonato, @dpiparo, @qliphy you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>