@slava77 2021-06-25
> enable gpu

@smuzaffar 
should I expect a "+1" from the bot to indicate that this hook was accepted and properly formatted, similar to what we get for the "test parameters"?