@fwyzard 2021-06-28
They restrict the scope of the atomic operations to a single CUDA block.

For "shared" memory this is always the case.

For "global" memory this may speed up the operation, for example by not
requiring to consider the cache of other blocks

See
https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#atomic-functions
for more details.