@slava77 2021-06-28
> Purely technical. no regression expected. no regression observed

there are quite a few plots in pixel tracking DQM which show small differences.
E.g. in 10824.502
<img width="742" alt="wf10824 502_pv_dz" src="https://user-images.githubusercontent.com/4676718/123711555-84e63100-d825-11eb-9084-7e181ac2bf1b.png">

I guess either the order of entries is not exactly the same or there are some differences in results at numerical precision level.
I suppose, these are expected.
Or is it worth to investigate in more detail?