@slava77 2021-06-29
> GPU tracking is known NOT to be fully reproducible for a variety of reasons that have been reported earlier in various meetings. (CPU tracking would also not be reproducible if the "order" of seeds/iterations whould change (say because of high granularity threading)

OK, thanks for clarifying. It was not particularly obvious that "no regression expected. no regression observed" was not a literal statement.