@slava77 2021-06-29
+reconstruction

for https://github.com/cms-sw/cmssw/pull/34250	48c6319
- code changes look reasonable
    - some summary of the functional changes in the PR description would be useful in addition to just providing evidence that this is better as measured
- jenkins tests pass and comparisons with the baseline show no (relevant) differences in the tested CPU workflows; the GPU workfow 10824.502 shows some differences in pixel track DQM distributions, mainly at numerical precision level, attributed to some changes in the order of the produced data.